transformer-waf/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiofiles>=23.0.0
watchdog>=3.0.0
loguru>=0.7.0
tqdm>=4.66.0
joblib>=1.3.0
pyyaml>=6.0
pydantic>=2.0.0
httpx>=0.25.0
redis>=5.0.0
requests>=2.31.0
